just using uint32 for masking buttons
git-svn-id: https://chan-sccp-b.svn.sourceforge.net/svnroot/chan-sccp-b/trunk@1981 fa19cda8-b258-4bce-aa1a-843e42da0b19

+++ src/sccp_mwi.c
@@ -404,8 +404,8 @@ void sccp_mwi_setMWILineStatus(sccp_device_t * d, sccp_line_t * l)
 	sccp_moo_t * r;
 	int instance = 0;
 	uint8_t status = 0;
-	uint64_t mask;
-	uint64_t newState = 0;
+	uint32_t mask;
+	uint32_t newState = 0;
 
 	if (!d)
 		return;
